misc/ssh-programmer/requirements.txt
paramiko==2.9.2
keystone-engine==0.9.2
